+++ src/main/java/sam/SamException.java
@@ -1,3 +1,4 @@
+package sam;
 public class SamException extends Exception {
   public SamException(String message) {
     super(message);
