+++ src/main/java/sam/command/AddCommand.java
@@ -1,5 +1,25 @@
+package sam.command;
 import java.time.LocalDate;
 import java.util.Map;
+
+import sam.Ui;
+import sam.parser.Parser;
+import sam.parser.SamInvalidDateException;
+import sam.parser.SamInvalidIntException;
+import sam.parser.SamInvalidTaskException;
+import sam.parser.SamUnknownCommandException;
+import sam.storage.SamSaveFailedException;
+import sam.storage.Storage;
+import sam.task.Deadline;
+import sam.task.Event;
+import sam.task.SamMissingTaskArgException;
+import sam.task.SamMissingTaskException;
+import sam.task.SamMissingTaskTitleException;
+import sam.task.SamMissingTaskValueException;
+import sam.task.Task;
+import sam.task.TaskList;
+import sam.task.TaskType;
+import sam.task.ToDo;
 
 public class AddCommand extends Command {
   private TaskType taskType;
