+++ src/main/java/sam/parser/SamInvalidDateException.java
@@ -1,3 +1,7 @@
+package sam.parser;
+
+import sam.SamException;
+
 public class SamInvalidDateException extends SamException {
   public SamInvalidDateException() {
     super("Please write dates as 'd/M/yyyy'!");
