+++ src/main/java/sam/parser/SamInvalidIntException.java
@@ -1,3 +1,7 @@
+package sam.parser;
+
+import sam.SamException;
+
 public class SamInvalidIntException extends SamException {
   public SamInvalidIntException() {
     super("Oops, I was expecting an integer!");
