+++ src/main/java/sam/parser/SamInvalidTaskException.java
@@ -1,3 +1,7 @@
+package sam.parser;
+
+import sam.SamException;
+
 public class SamInvalidTaskException extends SamException {
   public SamInvalidTaskException() {
     super("Oops, that task does not exist!");
