+++ src/main/java/sam/storage/SamLoadFailedException.java
@@ -1,3 +1,7 @@
+package sam.storage;
+
+import sam.SamException;
+
 public class SamLoadFailedException extends SamException {
   public SamLoadFailedException() {
     super("Oh no, there was a problem loading your list!");
