+++ src/main/java/sam/task/Deadline.java
@@ -1,3 +1,4 @@
+package sam.task;
 import java.time.LocalDate;
 
 public class Deadline extends Task {
