+++ src/main/java/sam/task/Event.java
@@ -1,3 +1,4 @@
+package sam.task;
 import java.time.LocalDate;
 
 public class Event extends Task {
