+++ src/main/java/sam/task/SamMissingTaskArgException.java
@@ -1,3 +1,7 @@
+package sam.task;
+
+import sam.SamException;
+
 public class SamMissingTaskArgException extends SamException {
   public SamMissingTaskArgException() {
     super("Oops, you're missing an argument!");
