+++ src/main/java/sam/task/SamMissingTaskException.java
@@ -1,3 +1,7 @@
+package sam.task;
+
+import sam.SamException;
+
 public class SamMissingTaskException extends SamException {
   public SamMissingTaskException() {
     super("Oops, you forgot to specify a task!");
