+++ src/main/java/sam/task/SamMissingTaskTitleException.java
@@ -1,3 +1,7 @@
+package sam.task;
+
+import sam.SamException;
+
 public class SamMissingTaskTitleException extends SamException {
   public SamMissingTaskTitleException() {
     super("Oops, you forgot a title for your task!");
