+++ src/main/java/sam/task/TaskType.java
@@ -1,3 +1,4 @@
+package sam.task;
 public enum TaskType {
   TODO, EVENT, DEADLINE;
 }
